requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
tabulate==0.9.0
psycopg2-binary==2.9.9
streamlit
plotly
pandas